Workflow + Excel (tests)

+++ src/PM/WorkspaceBundle/Controller/WorkflowController.php
@@ -13,7 +13,7 @@ class WorkflowController extends Controller
         
         
         // Récupération des statuts
-        $status = $this->getDoctrine()->getManager()->getRepository('PMWorkspaceBundle:Status')->findAll();
+        $status = $this->getDoctrine()->getManager()->getRepository('PMWorkspaceBundle:Status')->findBy(array('deleted' => false));
         
         // Récupération des roles
         $roles = $this->getDoctrine()->getManager()->getRepository('PMWorkspaceBundle:Role')->findAll();
@@ -37,10 +37,8 @@ class WorkflowController extends Controller
        // On coche les workflows déjà persistés
         $originalWorkflowsIds = array();
         foreach($originalWorkflows as $ow){
-            foreach($ow->getWorkflowsRoles() as $wr){
-                $originalWorkflowsIds[] = $ow->getOldStatus()->getId().'_'.$ow->getNewStatus()->getId().'_'.$wr->getRole()->getId();
-                $form->get('status_'.$ow->getOldStatus()->getId().'_'.$ow->getNewStatus()->getId().'_'.$wr->getRole()->getId())->setData(true);
-            }
+                $originalWorkflowsIds[] = $ow->getOldStatus()->getId().'_'.$ow->getNewStatus()->getId().'_'.$ow->getRole()->getId();
+                $form->get('status_'.$ow->getOldStatus()->getId().'_'.$ow->getNewStatus()->getId().'_'.$ow->getRole()->getId())->setData(true);
         }
        
         
@@ -52,26 +50,18 @@ class WorkflowController extends Controller
                     foreach($status as $statusTo){
                         if($statusFrom->getId() != $statusTo->getId()){
                             foreach($roles as $role){
-                                $workflowsRoles = array();
                                 if($form->get('status_'.$statusFrom->getId().'_'.$statusTo->getId().'_'.$role->getId())->getData() == true){
                                         if(!in_array($statusFrom->getId().'_'.$statusTo->getId().'_'.$role->getId(), $originalWorkflowsIds)){
-                                            // Créer le workflow s'il n'existe pas encore en BDD
-                                            //$workflow = repo
-                                            // si le nb res = 0 alors on créé 
                                             $workflow = new Workflow();
                                             $workflow->setOldStatus($statusFrom);
                                             $workflow->setNewStatus($statusTo);
-                                            
-                                            // Puis on ajoute le role
-                                            $wr = new WorkflowsRoles();
-                                            $wr->setRole($role);
-                                            $workflow->addWorkflowsRoles();
+                                            $workflow->setRole($role);
                                             $workflow->setWorkspace($workspace);
                                             $em->persist($workflow);
                                         }
                                 } else {
                                     foreach($originalWorkflows as $ow){
-                                        if($ow->getOldStatus()->getId() == $statusFrom->getId() && $ow->getNewStatus()->getId() == $statusTo->getId()){
+                                        if($ow->getOldStatus()->getId() == $statusFrom->getId() && $ow->getNewStatus()->getId() == $statusTo->getId() && $ow->getRole()->getId() == $role->getId()){
                                             $em->remove($ow);
                                         }
                                     }
@@ -83,7 +73,7 @@ class WorkflowController extends Controller
                 $em->flush();
                 
                 $this->get('session')->getFlashBag()->add('success', 'Workflow enregistré avec succès');
-                return $this->redirect($this->generateUrl('pm_role_index'));
+                return $this->redirect($this->generateUrl('pm_workflow_edit', array('id' => $workspace->getId())));
             }
         return $this->render('PMWorkspaceBundle:Workflow:form.html.twig', array('form' => $form->createView(), 'status' => $status, 'roles' => $roles, 'workspace' => $workspace));
     }
